must be long for zend_parse_parameters
git-svn-id: https://svn.php.net/repository/pecl/eio/trunk@331727 c90b9560-bf6c-de11-be94-00142212c4b1

+++ php_eio.c
@@ -1952,7 +1952,7 @@ PHP_FUNCTION(eio_readdir)
 {
 	char *path;
 	int path_len;
-	int flags;
+	long flags;
 	PHP_EIO_INIT;
 
 	if (zend_parse_parameters(ZEND_NUM_ARGS() TSRMLS_CC, "sllf|z!",
